# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SATBDS 2027 Symposium API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputCategory = typeof RegisterInputCategory[keyof typeof RegisterInputCategory];


export const RegisterInputCategory = {
  healthcare_professional: 'healthcare_professional',
  researcher: 'researcher',
  educator: 'educator',
  student: 'student',
  industry: 'industry',
} as const;

export interface RegisterInput {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  institution: string;
  country: string;
  category: RegisterInputCategory;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  attendee: 'attendee',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  category?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Speaker {
  id: number;
  name: string;
  country: string;
  /** @nullable */
  institution?: string | null;
  topic: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  initials?: string;
}

export interface SpeakerInput {
  name: string;
  country: string;
  institution?: string;
  topic: string;
  bio?: string;
  photoUrl?: string;
}

export type SessionSessionType = typeof SessionSessionType[keyof typeof SessionSessionType];


export const SessionSessionType = {
  keynote: 'keynote',
  panel: 'panel',
  workshop: 'workshop',
  oral: 'oral',
  poster: 'poster',
  opening: 'opening',
  closing: 'closing',
} as const;

export interface Session {
  id: number;
  title: string;
  day: number;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  room?: string | null;
  sessionType: SessionSessionType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  speakerId?: number | null;
  /** @nullable */
  speakerName?: string | null;
}

export type SessionInputSessionType = typeof SessionInputSessionType[keyof typeof SessionInputSessionType];


export const SessionInputSessionType = {
  keynote: 'keynote',
  panel: 'panel',
  workshop: 'workshop',
  oral: 'oral',
  poster: 'poster',
  opening: 'opening',
  closing: 'closing',
} as const;

export interface SessionInput {
  title: string;
  day: number;
  startTime: string;
  endTime?: string;
  room?: string;
  sessionType: SessionInputSessionType;
  description?: string;
  speakerId?: number;
}

export type SponsorTier = typeof SponsorTier[keyof typeof SponsorTier];


export const SponsorTier = {
  platinum: 'platinum',
  gold: 'gold',
  silver: 'silver',
  bronze: 'bronze',
} as const;

export interface Sponsor {
  id: number;
  name: string;
  tier: SponsorTier;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  website?: string | null;
}

export type SponsorInputTier = typeof SponsorInputTier[keyof typeof SponsorInputTier];


export const SponsorInputTier = {
  platinum: 'platinum',
  gold: 'gold',
  silver: 'silver',
  bronze: 'bronze',
} as const;

export interface SponsorInput {
  name: string;
  tier: SponsorInputTier;
  logoUrl?: string;
  website?: string;
}

export type RegistrationPaymentStatus = typeof RegistrationPaymentStatus[keyof typeof RegistrationPaymentStatus];


export const RegistrationPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  overdue: 'overdue',
  waived: 'waived',
} as const;

export interface Registration {
  id: number;
  userId: number;
  firstName?: string;
  lastName?: string;
  email?: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  country?: string | null;
  category: string;
  paymentStatus: RegistrationPaymentStatus;
  /** @nullable */
  paymentAmount?: number | null;
  registrationCode?: string;
  createdAt: string;
}

export interface RegistrationInput {
  category: string;
  dietaryRequirements?: string;
  specialNeeds?: string;
}

export type RegistrationUpdatePaymentStatus = typeof RegistrationUpdatePaymentStatus[keyof typeof RegistrationUpdatePaymentStatus];


export const RegistrationUpdatePaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  overdue: 'overdue',
  waived: 'waived',
} as const;

export interface RegistrationUpdate {
  paymentStatus?: RegistrationUpdatePaymentStatus;
  paymentAmount?: number;
}

export type AbstractAbstractType = typeof AbstractAbstractType[keyof typeof AbstractAbstractType];


export const AbstractAbstractType = {
  oral: 'oral',
  poster: 'poster',
} as const;

export type AbstractStatus = typeof AbstractStatus[keyof typeof AbstractStatus];


export const AbstractStatus = {
  submitted: 'submitted',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
  revision_requested: 'revision_requested',
} as const;

export interface Abstract {
  id: number;
  userId: number;
  /** @nullable */
  submitterName?: string | null;
  title: string;
  body?: string;
  abstractType: AbstractAbstractType;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  coAuthors?: string | null;
  status: AbstractStatus;
  /** @nullable */
  reviewNotes?: string | null;
  abstractCode?: string;
  createdAt: string;
}

export type AbstractInputAbstractType = typeof AbstractInputAbstractType[keyof typeof AbstractInputAbstractType];


export const AbstractInputAbstractType = {
  oral: 'oral',
  poster: 'poster',
} as const;

export interface AbstractInput {
  title: string;
  body: string;
  abstractType: AbstractInputAbstractType;
  keywords?: string;
  coAuthors?: string;
}

export type AbstractUpdateStatus = typeof AbstractUpdateStatus[keyof typeof AbstractUpdateStatus];


export const AbstractUpdateStatus = {
  submitted: 'submitted',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
  revision_requested: 'revision_requested',
} as const;

export interface AbstractUpdate {
  status?: AbstractUpdateStatus;
  reviewNotes?: string;
  title?: string;
  body?: string;
}

export interface Announcement {
  id: number;
  title: string;
  body: string;
  important?: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  body: string;
  important?: boolean;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface CountryCount {
  country: string;
  count: number;
}

export interface StatsSummary {
  totalRegistrations: number;
  totalAbstracts: number;
  totalSpeakers: number;
  totalSessions: number;
  pendingPayments: number;
  pendingAbstracts: number;
  registrationsByCategory?: CategoryCount[];
  registrationsByCountry?: CountryCount[];
}