# lib/api-zod/src/generated/types/abstract.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SATBDS 2027 Symposium API
 * OpenAPI spec version: 0.1.0
 */
import type { AbstractAbstractType } from './abstractAbstractType';
import type { AbstractStatus } from './abstractStatus';

export interface Abstract {
  id: number;
  userId: number;
  /** @nullable */
  submitterName?: string | null;
  title: string;
  body?: string;
  abstractType: AbstractAbstractType;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  coAuthors?: string | null;
  status: AbstractStatus;
  /** @nullable */
  reviewNotes?: string | null;
  abstractCode?: string;
  createdAt: string;
}